Network.AWS.CostAndUsageReport

Description

All public APIs for AWS Cost and Usage Report service

Service Configuration

costAndUsageReport :: Service

API version 2017-01-06 of the Amazon Cost and Usage Report Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by CostAndUsageReport.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when providing an invalid input. eg. Put a report preference with an invalid report name, or Delete a report preference with an empty report name.

InternalErrorException

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown on a known dependency failure.

DuplicateReportNameException

_DuplicateReportNameException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when putting a report preference with a name that already exists.

ReportLimitReachedException

_ReportLimitReachedException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when the number of report preference reaches max limit. The max number is 5.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

PutReportDefinition

module Network.AWS.CostAndUsageReport.PutReportDefinition

DeleteReportDefinition

module Network.AWS.CostAndUsageReport.DeleteReportDefinition

DescribeReportDefinitions (Paginated)

module Network.AWS.CostAndUsageReport.DescribeReportDefinitions

Types

AWSRegion

data AWSRegion

Region of customer S3 bucket.

Constructors

ApNortheast1

ApSoutheast1

ApSoutheast2

EuCentral1

EuWest1

UsEast1

UsWest1

UsWest2

AdditionalArtifact

data AdditionalArtifact

Enable support for Redshift and/or QuickSight.

Constructors

Quicksight

Redshift

CompressionFormat

data CompressionFormat

Preferred compression format for report.

Constructors

Gzip

Zip

ReportFormat

data ReportFormat

Preferred format for report.

Constructors

TextORcsv

SchemaElement

data SchemaElement

Preference of including Resource IDs. You can include additional details about individual resource IDs in your report.

Constructors

Resources

TimeUnit

data TimeUnit

The frequency on which report data are measured and displayed.

Constructors

Daily

Hourly

ReportDefinition

data ReportDefinition

The definition of AWS Cost and Usage Report. Customer can specify the report name, time unit, report format, compression format, S3 bucket and additional artifacts and schema elements in the definition.

See: reportDefinition smart constructor.

reportDefinition

Arguments

:: Text
rdReportName

-> TimeUnit
rdTimeUnit

-> ReportFormat
rdFormat

-> CompressionFormat
rdCompression

-> Text
rdS3Bucket

-> Text
rdS3Prefix

-> AWSRegion
rdS3Region

-> ReportDefinition

Creates a value of ReportDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rdAdditionalArtifacts - Undocumented member.

rdReportName - Undocumented member.

rdTimeUnit - Undocumented member.

rdFormat - Undocumented member.

rdCompression - Undocumented member.

rdAdditionalSchemaElements - Undocumented member.

rdS3Bucket - Undocumented member.

rdS3Prefix - Undocumented member.

rdS3Region - Undocumented member.

rdAdditionalArtifacts :: Lens' ReportDefinition [AdditionalArtifact]

Undocumented member.

rdReportName :: Lens' ReportDefinition Text

Undocumented member.

rdTimeUnit :: Lens' ReportDefinition TimeUnit

Undocumented member.

rdFormat :: Lens' ReportDefinition ReportFormat

Undocumented member.

rdCompression :: Lens' ReportDefinition CompressionFormat

Undocumented member.

rdAdditionalSchemaElements :: Lens' ReportDefinition [SchemaElement]

Undocumented member.

rdS3Bucket :: Lens' ReportDefinition Text

Undocumented member.

rdS3Prefix :: Lens' ReportDefinition Text

Undocumented member.

rdS3Region :: Lens' ReportDefinition AWSRegion

Undocumented member.