Network.AWS.CostAndUsageReport.DeleteReportDefinition

Description

Delete a specified report definition

Creating a Request

deleteReportDefinition :: DeleteReportDefinition

Creates a value of DeleteReportDefinition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdReportName - Undocumented member.

data DeleteReportDefinition

Request of DeleteReportDefinition

See: deleteReportDefinition smart constructor.

Request Lenses

drdReportName :: Lens' DeleteReportDefinition (Maybe Text)

Undocumented member.

Destructuring the Response

deleteReportDefinitionResponse

Arguments

:: Int
drsResponseStatus

-> DeleteReportDefinitionResponse

Creates a value of DeleteReportDefinitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsResponseMessage - Undocumented member.

drsResponseStatus - -- | The response status code.

data DeleteReportDefinitionResponse

Response of DeleteReportDefinition

See: deleteReportDefinitionResponse smart constructor.

Response Lenses

drsResponseMessage :: Lens' DeleteReportDefinitionResponse (Maybe Text)

Undocumented member.

drsResponseStatus :: Lens' DeleteReportDefinitionResponse Int

- | The response status code.