Network.AWS.CostAndUsageReport.DescribeReportDefinitions

Description

Describe a list of report definitions owned by the account

This operation returns paginated results.

Creating a Request

describeReportDefinitions :: DescribeReportDefinitions

Creates a value of DescribeReportDefinitions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdNextToken - Undocumented member.

drdMaxResults - Undocumented member.

data DescribeReportDefinitions

Request of DescribeReportDefinitions

See: describeReportDefinitions smart constructor.

Request Lenses

drdNextToken :: Lens' DescribeReportDefinitions (Maybe Text)

Undocumented member.

drdMaxResults :: Lens' DescribeReportDefinitions (Maybe Natural)

Undocumented member.

Destructuring the Response

describeReportDefinitionsResponse

Arguments

:: Int
drdrsResponseStatus

-> DescribeReportDefinitionsResponse

Creates a value of DescribeReportDefinitionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drdrsNextToken - Undocumented member.

drdrsReportDefinitions - Undocumented member.

drdrsResponseStatus - -- | The response status code.

data DescribeReportDefinitionsResponse

Response of DescribeReportDefinitions

See: describeReportDefinitionsResponse smart constructor.

Response Lenses

drdrsNextToken :: Lens' DescribeReportDefinitionsResponse (Maybe Text)

Undocumented member.

drdrsReportDefinitions :: Lens' DescribeReportDefinitionsResponse [ReportDefinition]

Undocumented member.

drdrsResponseStatus :: Lens' DescribeReportDefinitionsResponse Int

- | The response status code.